forms: rjsf: Extend ignoreEnterEvent for dynamically created text fields
